Create maintain_processes script
The IPFS daemon keeps shutting down unexpectedly.
As a workaround, I have created a script to periodically check
if the ipfs and node processes are alive, and restart them if not.

It prints the time whenever it restarts a process, so hopefully
that can be used to diagnose the issue.

Obviously, the output should be redirected a file so that we can
capture it.